Fix problem that intrinsic c_sizeof() "returns default integer instead of integer(kind=c_size_t)".
The symbol "c_size_t" is a Fortran PARAMETER defined in the iso_c_binding
intrinsic module; the same module defines the C_sizeof intrinsic.  The fix for
22315 does a look-up of "c_size_t" and uses its value to determine the return
type of the c_sizeof() intrinsic.  The problem reported here occurs when the
USE of the iso_c_bind module contains an ONLY clause that hides the c_size_t
symbol.  In this case, the c_size_t symbol is not defined.

Since the current value for c_size_t is 8 for all targets, this (less than ideal)
fix uses the value 8 to derive the type of the c_sizeof return value if the
c_size_t symbol is undefined.