Misc cleanup of the LLVM bridge
Add unreferenced variables to the debug information. These will
show up in the debugger as "optimized out".

Fix the scope of nested subroutines.

Clean up LL_DebugInfo so it is declared once.

Change Fortran symbols generated so CCSYMG is allowed. This is a
workaround as user-declared symbols are appearing in the bridge with
this flag set, but the flag should probably be cleared.

Misc. use of 'bool' type.