Add missing cases for finalization
//pgi/dev/comp/fe90/shared/src/semutil2.c

In gen_finalization_for_sym() - add third argument, memAst. This is
expected to be an A_MEM ast when set, otherwise 0. It allows us to
pass in derived type component expressions for finalization.

//pgi/dev/comp/fe90/shared/src/bblock.c

Add third argument to gen_finalization_for_sym() call.

//pgi/dev/comp/fe90/shared/src/semant.h

Update gen_finalization_for_sym() prototype with third memAst
argument.

//pgi/dev/comp/fe90/shared/src/semant3.c

In <assignment> ::= <psfunc> <var ref> <psfunc> = <expression>
production, add finalization to the LHS of the = if the LHS has
final subroutines.

//pgi/dev/comp/fe90/shared/src/semfunc.c

In func_call2() - if a function result has final subroutines, then
we need to finalize the function result after assigning it to the
LHS of the function expression. We also apply this same check in
ptrfunc_call() (for procedure pointers).