Fix for missing dtype for symbol in ILM file
The problem is that a symbol is being placed in th ILM file
and its data type is not.  The symbol in question, a dummy
argument, and its owning subprogram should not be lowered
(and placed into the ILM file) because the subprogram is not
used. A previous change is causing these symbols to be lowered.
Backing out this fix does not seem to cause any problems
(even for prior regression tests).

Also deleted a dangling sentence fragment from the symtab
documentation.