Improve handling of pointer functions with array return types
For pointer function with adjustable array or any array as
return type, we should use the return type of the function
based on its interface. The caller passes in the size but the
callee will create an adjustable array based on that size.
Also when we create temporary array for return type, we should
clear PARAMCT field because it is overloaded field and SOCPTR
also uses that field.