Better handling of function via pointer with adjustable return type
When function is called via pointer and the return type is
adjustable, make sure to call ref_entry(fval/iface) so that
we don't try to allocate an adjustable length as a result at
the call site when we don't need to.  Also clear SOCPTR field
in gen_pointer_result and gen_allocatable_result.