Fix a long-standing subtle bug in BIH handling
addnewbih(after,,) allocates a new BIH and inserts it into the list of BIHs after 'after'
and fills in BIH_PREV and BIH_NEXT fields as necessary.

By convention, BIH_PREV(0) points to the last BIH in that list, and BIH_NEXT(0) points
to the first, and there are places in the compiler that walk the BIH list forwards
or backwards from BIH_NEXT(0) or BIH_PREV(0).

In addnewbih(), if BIH_NEXT(after) is negative, then don't link BIH_NEXT(newbih)
and don't link BIH_PREV(BIH_NEXT(after)).