Fixes problem with llvm terminated by signal 11
See fix_symbol(). Also some code cleanup.

Not all symbols in the table were being properly rewritten, so if
the ones not updated were used, the compiler could crash with a
segfault addressing unmapped memory off the end of the symbol table.
The fix was to process all the symbols in the table by changing
the loop to walk over each entry in the table.