Additional work on proper FINALIZATION
In <psfunc> <var ref> <psfunc> = <expression> production of semant3.c,
do not perform the finalization of the LHS if it's allocatable and
we're using F2003 allocation semantics or if the LHS is a pointer. If
we're using F2003 allocation semantics, then the LHS will get
finalized during (automatic) deallocation. If the LHS is a ponter,
then the language spec allows us to not finalize them (it's
processor dependent).

In func_call2() and ptrfunc_call() of semfunc.c, Need to finalize the
function result if it's not allocatable or pointer. If it's
allocatable, then finalization occurs as part of (automatic)
deallocation. If it's  pointer, then the language spec allows us to not
finalize it (it's processor dependent).

In gen_finalization_for_sym() of semutil2.c, process the array case
regardless of whether it has a descriptor or not.

In rewrite_allocatable_assignment() of transfrm.c, add calls to
finalize() in various deallocate cases. When assigning a scalar to
an allocatable array and we're using F2003 allocation semantics,
allocate the array with a default size first. If the array is
already allocated, then finalize the array (if applicable) before
assigning to it.