Remove a check of ADDRTKNG from IL_LDA
It is unclear why the check is there.   However, the flag is
only an indication that the address has been taken and should
not be used to generate the load.

--
May need to go back and do for c/c+ it does not looks right either but will need to find an example for it.