Fix issue with polymorphic select type
In transform_call(), fix condition that checked an actual
argument to see if we need to assign its type to its type
descriptor when the object is non-polymorphic. We need to
perform this assignment for objects with regular descriptors
as well as extended scalar descriptors. We do not want to
perform this assignment for objects that have a static type
descriptor since it already represents the type.