Fix handling of alloca() in LLVM bridge
Calls to alloca() and __builtin_alloca() were not being handled
correctly on x86.  They were not handled correctly on Power if
the alloca() call was the outermost expression in an expression
statement (which is useless, but legal).

Fix targets (other than Power, which was already correct)
to not convert __builtin_alloca() calls into alloca() calls.
The back ends look for __builtin_alloca() to do their special
processing of using the "alloca" instruction.

Fix targets to check for __builtin_alloca() calls when
processing an expression statement.  This check was already
happening for expressions in other contexts.  It was just
missing for expression statements.