Fix bug in name resolution of type-bound procedure.
Name resolution was going wrong when the same name was used for a
binding-name and a procedure-name. Specifically, we were getting "Interface
is not compatible with parent's interface for type bound procedure bad" in a
case like this:
  type :: type1
  contains
    procedure :: bad => foo
  end type
  ...
  type, abstract, extends(type1) :: type2
  contains
    procedure :: bar => bad
  end type
  ...
  type, extends(type2) :: type3
  end type

The problem is that the implementation name was being passed to
get_implementation when it was expecting a binding name.