Fix erroneous export of procedure binding name.
Example fragment of module:

  type :: T1
  contains
	procedure :: bad => subr1
  end type
  type :: T2
  contains
	procedure :: bad => subr2
  end type

If only T1 is present the name "bad" is not exported from this module but
when T2 is present it is (but shouldn't be).

When the first procedure binding is encountered, and new symbol is create
for "bad". Its stype is ST_UNKNOWN, not ST_PROC, so IGNORE is set on it.
Then a ST_PROC symbol is created from that, also with IGNORE set.

When the second binding is encountered, we find the ST_PROC symbol for bad.
A new symbol is created from that, but we fail to set IGNORE, so it gets
exported.

The fix is to copy the IGNORE flag to the new symbol to prevent the export.