Disable collapsing of distributed parallel loops.
Collapsing is disabled due to a performance bug when the array expression is
complex. We may be able to mark the variable so that the optimizer can do
better job. Performance is not specific to distributed loops; this also
happens in normal parallel loops too.

Also make sure all sequential loop index variables are private if they are not
already private.