Propagate dtype from procedure pointer's interface.
In <primary> ::= <var ref> of semant2.c, if <var ref>
is a procedure pointer, then we need to propagate the dtype from the
procedure pointer's interface if it's a function.

Also added ast_is_sym() utility function to determine if an AST
is suitable for a call to sym_of_ast() and memsym_of_ast().