Fix potential name clash in type descriptor name.
In sym_get_sdescr() of rte.c, we typically create a type
descriptor symbol name by fusing the module name and type name together
with an underscore separating the two names (i.e., moduleName_typeName).
See the source code for a couple other variants. In all cases each
portion of the name is separated with an underscore. Because a user
can include an underscore in type and module names, it's possible to
construct an example that causes a naming conflict. To correct
this problem, we append a number of underscores equal to the number
specified by the user to the module and type name. This prevents the
user from constructing a failing case.

Also added the support functions, lenWithUnderscores() and
mangleUnderscores() to aid in constructing the mangled name.