Back out change to symbol reading.
The problem with the earlier change is that the loop adds new symbols to the
end of the symbol table, and it is critical that those new symbols are not
processed by the loop.

The change loops over the symbol table and processes all the
symbols. The old algorithm tries a much wider range of integers,
tests if they are in a map, and then processes the symbol. This
may fail to process some symbols and is slower.