Fix two Fortran features for an important test case: (1) CHARACTER actual arguments passed for scalar unlimited     polymorphic (CLASS(*)) dummy arguments need to convey     the values of their length type parameters, and (2) deferred-length CHARACTER array assignments.
This patch extracts and abstracts some code that was written
for a similar task during the implementation of construct associations
in semant3.c, packaging the code as new utilities in symutl.c, and
calling these new functions from the spot in rest.c that handles the
case of CHARACTER actual argument association with an unlimited
polymorphic (CLASS(*)) scalar dummy argument.

The other changes in symutl.c add support for deferred-length
CHARACTER array temporaries.

In semfin.c, the type used to describe the type descriptor
additional argument created as a special case for a scalar
polymorphic dummy is changed from a simple int to an array,
so that the length can be extracted in flang2 correctly;
it previously elicited a assert in DEBUG code, or led to
bad code otherwise.

The changes in func.c are cosmetic improvements that were
made while debugging and seemed worth applying.