Replace GENERIC_TBP_ONCE macro with function.
In queue_generic_tbp_once(), add the check that was originally in
GENERIC_TBP_ONCE macro and also check whether we are in a nested subprogram.
If we are, then we also need to regenerate the generic type-bound procedure
bindings once.

Before, we were never regenerating them because they were already
generated by the host subprogram.