Create CSE for atomic cmpxchg instruction.
The atomic cmpxchg instruction is referenced/used by both IL_CMPXCHG_SUCCESS
and IL_CMPXCHG_OLDx. We need to create a CSE for it so that the instruction
does not get generated twice.

Remove a check in cgmain which try to reuse tmps if it already exists for
cmpxchg ili. That is not a good solution as we may use the same ili in other
contexts.