Problem finding descriptor for move_alloc call.
When get_member_descriptor() fails to find the descriptor we should fall back
to the SDSC. Simplify the logic by using find_member_descriptor().

Extract the common code for finding the descriptor into find_descriptor()
so it can be used for both args.

get_member_descriptor() was sometimes returning the next member of the derived
type. Reimplement it to go through the compiler-generate members after the
real member and only return one that has DESCARRAY set.

Add a check for valid sptr in lower_sptr() so when it fails it is more
apparent what is going wrong.