Fix a bug with OpenMP parallel workshare directive.
In opt_allocate() of commopt.c, we may replace temporary array bounds
created by the compiler with actual bounds that are used when we allocate
the array. When this happens, disable the allocate optimization.
Otherwise, in rewrite_all_shape(), do a check and set parref on the new
symbol table pointer (sptr). This should only occur when STD_PAR is not set.

When converting forall workshare to a parallel do in convert_omp_workshare()
(in outconv.c), we need to make sure to increment/decrement the dolevel
variable every time we encounter "do" and "end do". This ensures that "pdo" and
"end pdo" asts are emitted properly.

In rewrite_all_shape() of commopt.c, store the results of ast_rewrite() before
assigning it to SHD_xxx structures because there may have been a reallocation
of the shape ast in one or more calls to ast_rewrite().
Also change the dimemsion size in the for loop from a hardcoded 7 to MAXDIMS.

In set_parref_flag() function of semsmp.c, do not add the variable to uplevel
struct if it is a member.

In rewrite_all_shape() function of semsmp.c, only set PARREF (i.e., put in
uplevel struct) if new_sptr != old_sptr.

In set_parref_flag2() function of semsmp.c, if new_sptr is global/static, do
not put new_sptr in the uplevel struct.