Fix a bug with procedure pointers of functions with a hidden argument.
Put/resolve hidden argument of a function pointer which we accessed through
the interface field of its dtype.  This is needed because in the
transform_call() function, we expect that it has already been resolved.
Otherwise, the number of arguments for the procedure call is incorrect and
the compiler may access invalid memory when it loops over the
dpdsc (argument) structure.

Also undo hidden arguments for internal subroutines in the fix_symtab()
function.

Add an assert in the transform_call() function that checks whether we have
resolved a hidden argument.