Improve module file consistency checks.
For now, this check is disabled using an ifdef macro (DO_MODULE_OPTION_CHECK).
The goal of the consistency check is to make sure we are not mixing modules
that are compiled with and without certain compiler flags like
-Mlarge_arrays.

The consistency checks do not yet work with libraries that are portable
across all compiler flag-sets. So, the checks are disabled for now.