Don't ignore 64-bit division by constant 2^32
In the addarth() function of the iliutil.c source file:

When performing a 64-bit integer division, either signed or unsigned,
with a denominator known at compiler time, the compiler checks if the
denominator is "1", in which case the division can be optimized away
entirely.  But when checking for a value of "1", the code was mixing up
the high-order and low-order 32 bits, so the code was thinking that a
denominator of 0x100000000 (2^32) was "1" and was incorrectly getting
rid of the division in that case.