Add more infrastructure for AVX-512 support in machreg.h.
Made the following additions to support the implementation of AVX-512
'opmask' registers:

-- Declared a new enumeration containing symbolic names for the 8
opmask registers, namely OR_K0, OR_K1, ..., OR_K7.

-- Added a new macro 'N_OPMASK_REGS' which is '#define'd to the total
number of opmask registers, namely 8.

-- Declared a new array "char *opmask_reg[N_OPMASK_REGS + 1]" which
will contain the assembly code representations of the 8 opmask
registers, and a new macro 'OPMASK_REG_NAMES' which is '#define'd to
the list of initialisers for this array.