Fix a bug with overloading of type bound procedures.
Call findByNameStypeScope() and pass the scope to a functon so that it returns the
correct tbp implementaion routine/function as there can be multiple functions with
the same name in different scopes.  The getsymbol() function can return any symbol
that is the top of hash table which may not be in the scope we want.  If
getsymbol() returns an inappropriate symbol, the symbol may not have its
paramaters fixed.  Note that only routines that are referenced will have
arguments/descriptors/etc.