Fix error with calling an adjustable array-valued internal function.
When calling an adjustable array-valued internal procedure, a temporary
representing the result of the function is allocated in the caller. Code
is generated to compute the size of the temporary and is incorrect in the
following situation:
- the call occurs in an internal procedure
- the host of the caller (& callee) is a module routine
- the size of the result depends on a dummy argument of the host
Because the host is a module routine, there are two sym entries for the host:
1) an ST_PROC which represents the host's 'interface'
2) an ST_ENTRY which is used the body of host and its internal procedures
   are compiled.
Each of sym entries has its own set of symbols representing the dummy arguments
of host.

When the array-valued internal procedure is compiled for the first pass, its
host context is the ST_PROC; thus references to the host's dummies will be
resolved by the ST_PROC's set of symbols.  When the caller of the array-valued
function is compiled (occurs in the second pass), the host scope will be the
host's ST_ENTRY sym entry. Thus, any reference to a host dummy will be resolved
by the ST_ENTRYs set of symbols.  If the size computation of the array temporary
references a host dummy, it's resolved by the ST_PROC's set rather than the
ST_ENTRY's set.

The solution is replace the ASTs of the ST_PROC's dummies with the AST's
ST_ENTRY's dummies and then perform an ast rewrite of the computation.
NOTE that there is already a rewrite process in place because we first
substitute the ASTs of the callee's dummy arguments with the ASTs of the
actual arguments.

symutl.c - add utility function get_symtype()
Searches the remaining hash list of symbol looking for a specific name
and SYMTYPE.

symutl.h - prototype for get_symtype()

semant.h - Add sem.modhost_proc & sem.host_entry:
ST_PROC & ST_ENTRY, resp., of a module host routine containing an internal
procedure.  These members will be set & referenced when an array valued
function is processed in semfunc.c:gen_array_result().

semant.c - initialize sem.modhost_proc and
sem.modhost_entry = 0;

semfunc.c
func_call2() - a few cosmetic changes - change a couple of compound statements
to simple sequence of statements; decl's were absent (suspect that at one time
they were guarded by a conditional

gen_array_result():
- add a new argument (callee) to the end; it will be the sptr of the
  function being called. It's needed so the gen_array_result() can determine
  if the callee is an internal procedure.
- as a part of generating the assignments to the bound temps of the array
temp, determine if the calling context is an internal procedure calling an
internal procedure and the host is a module routine.  If so, identify the
host's ST_PROC & ST_ENTRY; if present, walk the param dpdscs of the two
entries and replace the ID asts of the ST_PROC dummies with the ID asts
of the ST_ENTRY dummies.