Add bug fixes for pointer assignment, sourced allocation, etc.
In check_pointer_type() of func.c, set the pointer's type
descriptor to its declared type when the pointer target is
null().

In transform_call() of transform.c, fix some code that
generates calls to set_type for descriptor arguments. Also
fix the generation of descriptor code for ST_MEMBERs in
which we are using a local descriptor (i.e., a surrogate descriptor)
for the actual argument. We need to make sure that the surrogate is
locally declared.

In sourced_alloc_and_assign() of type.c, we want to recursively walk
the layout descriptor if a layout descriptor's tag is 'F' which
represents a non-pointer/non-allocatable finalized component
(e.g., a component that is a finalized derived type). Also
return after printing "__NONE" and "unknown" baseTags in
__fort_dump_type(). Otherwise, we might get a seg fault when we
try to display additional information about the type descriptor.

In resolveBind() of semtbp.c, remove incorrect check on
curr->isDeferred.

Reformatted some code to LLVM standards.