Fix a bug with optional dummy arguments.
Avoid a piece of special-case processing for optional
assumed-shape array dummy arguments (or contiguous sections
of the leading dimension of such) being passed as actual
array arguments when the optional dummy argument is assumed-size
character.  The special processing does not appear to be needed
in that case, and the assumed character length is being lost.
(Also, fix some spelling errors I noticed during debugging.)