Fix spurious 'must be a function name' error.
When we encounter ENDFUNCTION or ENDSUBROUTINE in an INTERFACE, do_iface_module()
was not always finding the interface correctly. The problem is in the call chain:
do_iface_module() -> _do_iface() -> refsym() -> sym_in_scope(). In the example
below, iface2 is still on the scope stack when we search for iface1 and that
causes it not to be found. A new symbol is created for iface1 which does not
have the right type.

The solution is to pop the scope stack for iface2 before calling do_iface_module()
so that iface1 can be found correctly.

In the ENDFUNCTION case there was an extra call to do_iface_module() that was
removed.

The changes to _do_iface() are just cleanup, no function changes intended.
The logic was simplified and int types where changed to enums where possible.

Test case:

module m
  implicit none

  abstract interface
    subroutine iface1()
    end subroutine
  end interface
  procedure(iface1), pointer :: registered_proc => null()

  abstract interface
    subroutine iface2()
    end subroutine
  end interface

contains

  subroutine register(proc)
    procedure(iface1) :: proc
  end subroutine

  subroutine sub()
  end subroutine

  subroutine test()
    ! Incorrect error occurs here:
    ! Argument number 1 to register must be a function name
    call register(sub)
  end subroutine

end module