Set correct array bounds on allocatable assignments.
When assigning an allocatable array with F2003 allocatation semantics, we set the
bounds to 1:extent in the case of a strided section but it should happen
with any array section.

The fix is to change mk_bounds_shape to always make the bounds 1:extent
and call it whenever the bounds need to be fixed up. This makes
the function all_stride_one_shape() unnecessary.