Fix an internal compiler error (ICE) when a formal is CLASS(*) and the actual is a function call.
In this case we need to pass a descriptor arg but end up using symbol 0,
leading to the ICE.

The fix is to create a temporary for the descriptor. The code from the
unlimited polymorphic case of A_BINOP, A_PAREN, etc. has been extracted
into a new function, temp_type_descriptor(), so that it can be used
in the A_FUNC case. The code around it has been reorganized.

The special case for elemental_func_call() has been removed because
the general case seems to handle it correctly.