Support MERGE in data initialization.
Add AC_I_merge to enumerations in both versions of semant.h.
Add versions of eval_merge() in semutil2.c and dinit.c.
Add case for MERGE in eval_init_op(), construct_intrinsic_acl(), eval_init_op().

In get_ac_intrinsic(), allow for getting the I_* constant from A_OPTYPE of
the ast when it is an intrinsic function call. This is necessary because
MERGE has already been rewritten as an ST_PROC (e.g. a call to mergei()).
To do this, extract the switches over PD_* and I_* values into separate
functions so that map_I_to_AC() can be re-used in the ST_PROC case.

Make this work for I_MIN and I_MAX as well by having ref_intrin() set A_OPTYPE
in the call AST to I_MIN/I_MAX the same way it is done for other calls.
This makes the strcmp kludge for MIN and MAX in construct_intrinsic_acl()
unnecessary.

The rest of the changes in semfunc.c are just cleanup.