Fix return value of initFinalSub().
In initFinalSub() of semtbp.c, we need to return -1 if a final subroutine is
defined in a different scope, we are in the module specification section
(i.e., we have seen the "contains" statement but not the "end module"
statement), and we are in the first semantic pass. In other words, we have not
seen the entire module yet so we cannot safely proceed. Otherwise, we could
have incomplete information like in the case of overloading the name of the
final subroutine.