-O2 gives wrong results
A contiguous pointer section is passed to a routine whose corresponding dummy
is assumed-shape.  The callee expects that the passed address to locate the
first element of the actual argument. However, when a target section is
pointer assigned to a pointer, the pointer can locate the beginning of the
target array rather than the start of the section. For example, given
  ptr => target(2:30)
ptr is assigned the address of target and lbase in the ptr's descriptor is
assigned a non-zero offset so that that adjusting ptr with lbase is the
address of target(2).

The 'pointsto' analysis/optimization is used in the compiler to track pointer
assigements of contiguous (stride 1)  sections.  The bug has to do with the
whole section of the pointer being passed to a routine whose corresponding
dummy is assumed-shape.  In rest.c, stride_1_is_section() returns TRUE given
the pointsto info. transform_section_arg() then elides the 'copy_f90_argl'
calls for the pointer section argument.  Without the copy call, the address
passed is not the address of the first element.

rest.c:
- stride_1_section(). Add arguments entry and pos.  entry is the sptr of the
callee and pos indicates which argument.  With the new arguments,
stride_1_section() can examine the attributes of the dummy corresponding to
the actual when a pointer section. If assumed-shape, FALSE must be returned.
- cleanup. tabs removed; indentation