llassem.c cleanup
Rename local variable gblsym to to vargblsym and typegblsym to reduce ambiguity.

Normally AG_DSIZE will result in actual size but for accdatainit, size is not
available.  Once the data initialize is done for extern_init, mark AG_DSIZE = 1:
this mark that the data initialize is done so that we don't emit duplicate
init code.