Initial check-in for enhanced error message support.
This adds column number and error location (i.e., displays location with a ^)
to error messages. The old style error messages are currently the default.
Compile with -Hx,1,1 to get "enhanced syntax error" messages. Semantic errors
require a change to the call site. There is a new function called
errWithSrc() that takes the same arguments as error(), but also requires
three additional arguments to properly display the column number/location.
Additional documentation on the calling convention will be available soon.

For now, note that this initial check-in provides enhanced error message
support for front-end syntax errors and semantic errors that occur during
parsing. Enahnced error support for lowering and the back-end TBD.