Finish implementation of calling our complex math routines.
For passing & returning scalar complex, we want to follow the C ABI. The
operands & results of complex ILI are presented as packed (combined);
when we call our routines, we can use either the C ABI or the vector ABI
where a scalar complex is viewed as a complex vector of length 1 (or
'packed').

The 'packed' scheme was already used for divide in the work to present
complex as a first class operation.  In the new math naming scheme, we are
now calling routines using the C ABI.  Generally, there is no difference
between the C ABI and vector ABI.  For double complex on x64, the C ABI
says to view complex as a struct of 2 components; when passed & return,
two xmm's are used.  In the vector ABI, the complex is packed into a single
xmm.

NOTE that this work requires the RTE which uses the C ABI for the _1 routines
and the presence of the _1v routines for complex double.