Fix bug with call to function pointer returning allocatable array.
When we make space in the DPDSC (dummy parameter descriptor) of the dtype
for a procedure for the return value of a function pointer, we need to add
the return value symbol. If we don't, it is not available to add the array
descriptor to the argument list in some cases. Then it is omitted when the
function is called, even though the callee expects it.

Fixes #229