Fix how type-bound procedure forward references are handled.
Setting the HIDDEN flag for a type-bound procedure causes a problem for
the length function when the compiler tries to call queue_tbp with
task=TBP_COMPLETE_FIN in semfin for following code snippet.

Example 1:

module x
  type, abstract, public::xx
   private
   contains
    procedure::length
  end type xx
 contains
 integer function length(this)
 ...
 end function length
end module x

The HIDDEN flag was set to fix this example:

Example2:

module xxx
 type,abstract,public:: type1
  contains
   procedure(ds_oprnd_sync_i),deferred,public::sync !line 4
 end type1

 type, abstract, public:: type2
  procedure, public:: mark_undelivered=>suba
  contains
  ...
 end type type2

 private suba
 subroutine suba(this,ierr,sync)  !line 14
 ....
 end subroutine

end module xxx

The problem in example 2 is that when compiler creates the symbol for the type
bound procedure sync at line 4, it has STYPE as unknown, let says it create
symbol number:1234.

When we get to line 14, compiler calls getsymbol which returns symbol:1234
because symbol:1234 has STYPE as unknown.  This should raise a flag that if we
are going to use the symbol for a dummy argument.  Basically, the dummy
argument symbol number would be created after the routine symbol.  It then uses
that symbol as dummy argument for suba. We have not experienced any problem so
far.

Some time later when we restore host_state, symbol:1234 is reversed to its
previous state(module state). Its storage class, OPTARG, and its dtype are
gone. This result in a complain in get_kwd_args.

The solution is not to set HIDDEN, but then at the point where we set up the
dummy argument we check if its argument symbol number is less than its own
routine, and get a new symbol.