Fix bug with unlimited polymorphic pointer assignment.
When pointing an unlimited polymorphic pointer at an unlimited polymorphic
dummy argument, we were failing to copy the length in the case that the
actual argument is a character string. The fix is to change RTE_set_type
(and RTE_test_and_set_type) to copy the length as well as the type in this
case.