Fix ICE: mk_mem_ptr_shape: lwb not sdsc
The problem was the calls to fix_mem_bounds() in transform_move_alloc(). They
modified the array descriptor (ADSC) corresponding to the dtype of member 's'
of one derived type (IntegerStackType) and that affected the dtype of member
's' of another derived type (IntegerStackStackType).

The fix is to instead pass the parents (of "from" and "to") in to
copy_surrogate_to_bnds_vars() and change that to add the parent (with
add_parent_to_bounds()) when the bounds are accessed. This requires
add_parent_to_bounds() to allow for parent == 0.

Additional cleanups:
- copy_surrogate_to_bnds_vars() only needs the dtypes of the source and
  destinations, not the symbols themselves, so just pass those in
- add move_alloc_arg() to encapsulate the processing done on both args

Fixes #196