Fix alignment of stack variables.
The matrix multiply routines in the runtime require stack variables
to be 16-byte aligned, at least for x86. The are assembly language
implementations that depend on that to be able to use SSE2 instructions.

align_of_var() was changed to add the flg.quad check. It was moved to
llassem.c because that's where MIN_ALIGN_SIZE and DATA_ALIGN are defined.

data_align is only ever set to DATA_ALIGN or 15, so eliminate it and use
DATA_ALIGN instead.