Fix for: Lowering Error: allocate with no array size
The error occurs in this code fragment:
	integer :: x(2:1)
	integer, allocatable :: y(:)
	allocate(y, source=x)
We try to get NUMELM from the dtype of x which has it set to AST 0.
That leads to the internal error in lower_numelm.

The fix is to set NUMELM to astb.bnd.zero instead of AST 0 in
fill_fixed_array_dtype() when we see that it is a zero-size array.