Fix compile-time infinite loop.
During import, do not set SLNK field to 0 since the aux.lists that
hold these are terminated with NOSYM. Instead, try to remap
the symbol in the SLNK field or leave it alone. We assume that
a symbol's SLNK is 0 when the symbol is not in an aux.list.

Also added some "debug" assertions to make sure we do not corrupt the
aux.lists.