Fix verifier error with large arrays.
We are expecting an ILIA_IR, but we are having an ILIA_AR. We do ITOI8 to PLD and
BASE address, which are not correct. Fix is when we check subscript, when it's
an array, do not do the convert yet. Correct conversion will be handled later
in lowerexp.c.