Fix ICE related to CONTIGUOUS attribute
The problem only happens when we pass in the result of fast matmul of
CONTIGUOUS matrix as an argument to a function whose return value is an array
whose size depends on the argument. In this case, a temporary is created, so
matmul_use_lhs returns TRUE, which causes mk_matmul_sptr is not called,
therefore NOEXTENT field is not set when we check it in mmul_arg. Add another
check besides NOEXTENT so that we can compute the correct ldim.