Fix ICE: check_member: cannot match member with derived type
In inline_reduction_f90(), call check_member() with dest's parent ast when dest
is a member of another member of a derived type. That is, d%m1%m2 where d is a
derived type object, m1 is a member of d and m2 is a member of m1.

Otherwise, the do_check_member_id() function will check the wrong container
dtypes of dest and ast2.