Further support for conditional vectorization.
cgmain.c:
If masks are generated for conditional vector code, must make mask size to
match computational size if an intrinsic math call is made. But blending
values at the end of a conditional LLVM uses bit-vector masks with the
SELECT instruction.

iliutil.c:
Fix call to mk_prototype so that the potential mask argument to math
instrinsic calls have the right type and size.