Fix use of uninitialized variable in chk_struct_constructor().
In some cases of the switch in chk_struct_constructor(), ast does not get
set. But it is used in the condition to see if "Illegal expression in
initialization" error should be emitted.

The fix is to initialize ast to 0 and check for that case, assuming that
ast = 0 means the error should be emitted. This makes it even more
complicated, so it was extracted into is_illegal_expr_in_init().

This is the test case:
  type t
    integer, allocatable :: x(:)
  end type
  type(t) :: y = t([1, 2, 3])