Fix ICE: remove_stmt: corrupt STD or deleting statement twice
In forall_dependency(), the calls to rdilts() and wrilts() are causing the STD
list to be corrupted.

rdilts() saves in rw_state.std_{before,after} the STDs before and after the new
flow graph node, in this case the prev and next of newstd. Then in wrilts() those
to STDs are linked together in the list, leaving out newstd.

As a result, newstd is removed incorrectly and the doubly linked list of STDs
is messed up: STD_NEXT(STD_PREV(newstd)) != newstd.

One manifestation of this bug occurs if we later attempt to remove newstd. We
notice the corrupted list, leading to this ICE: "remove_stmt: corrupt STD or
deleting statement twice".

This is the test case:
real, pointer :: x(:), y(:)
y = y + 1
where (x /= 0.0) y = y + x

Some unused locals were also cleaned up.