Make begin_cal/add_arg more robust
begin_call and add_arg are used to build a function or subroutine call
AST node and to build the argument list.  Unfortunately, it doesn't support
nesting, meaning you can't begin to build another procedure call tree
before finishing the previous one.  I ran into several cases which had
  begin_call(A_CALL, func, 3);
  add_arg(ast1);
  add_arg(ast2);
  i = somefunction(...);
  add_arg(i);
where 'somefunction' started another begin_call sequence.  This overwrote
the curr_call fields, so the last add_arg was adding an argument to the
second call, invalidly, and not writing the last argument to this call.
So, do some error checking here, that we aren't writing to an argument
number greater than the number of arguments that we allocated.
Also unfortunately, there's no function call to check that all the
arguments are written.

In begin_call, fill in the new field curr_call.arg_count with the number
of arguments that this function wants.  In add_arg, check that the
curr_call.arg_num doesn't go past curr_call.arg_count.  In the example
above, the last add_arg would look like it's adding an extra argument
to the procedure call created by somefunction(), and would overrun that
arg_count, and so create the error.  I thought about adding a test in
begin_call to check that the previous procedure call actually completed
(arg_num == arg_count), and I'll try that out next.