Fix internal compiler error.
Need to reset SYMLKP(gbl.entries) = NOSYM once it is done
emitting the code because gbl.entries does not get set
until new routine to new routine is processed and outlined
function may try to do emit entries if we don't reset it.
Also check if master_sptr == 0, before trying to emit entries.
Also don't check the argument number for function with
entries when we try to emit the jump instructions because
it can have arguments.